Extract CreateRoomForm utility functions
