+++ src/pages/AdminDashboard/CreateRoomForm/CreateRoomForm.tsx
@@ -28,8 +28,17 @@ import {
 import Flags from 'country-flag-icons/react/3x2'
 import { parsePhoneNumber } from 'libphonenumber-js'
 import { useTranslation } from 'react-i18next'
-import NotificationSelector, { UNITS } from './Reminders/NotificationSelector'
-import { validationSchema } from './formValidation'
+import NotificationSelector, { UNITS } from '../Reminders/NotificationSelector'
+import { validationSchema } from '../formValidation'
+import {
+    mapDestinationsToInputField,
+    getRemindAt,
+    validateEmail,
+    validatePhoneNumber,
+    isParticipantValid,
+    isNumeric,
+    formatDestinations,
+} from './createRoomTools'
 
 export interface Room {
     id: string
@@ -47,16 +56,6 @@ interface Notification {
     number: number
 }
 
-const mapDestinationsToInputField = (destinations) => {
-    const values = [] as any
-    for (const destination of destinations) {
-        const [value] = Object.values(destination)
-        values.push(value)
-    }
-
-    return values
-}
-
 const Button = styled(MuiButton)(spacing)
 
 const CreateRoomForm = ({ fields, onFormSubmit, onCreateFormReset }) => {
@@ -67,23 +66,6 @@ const CreateRoomForm = ({ fields, onFormSubmit, onCreateFormReset }) => {
         unit: UNITS.mins as Unit,
         number: 0,
     })
-
-    const getRemindAt = (startAt, notification) => {
-        let remindBeforeTimestampSecs = 0
-        switch (notification.unit) {
-            case UNITS.mins:
-                remindBeforeTimestampSecs = notification.number * 60
-                break
-            case UNITS.hours:
-                remindBeforeTimestampSecs = notification.number * 60 * 60
-                break
-            case UNITS.days:
-                remindBeforeTimestampSecs = notification.number * 60 * 60 * 24
-                break
-        }
-
-        return startAt - remindBeforeTimestampSecs
-    }
 
     // Populate the form with fields values when isEditing
     useEffect(() => {
@@ -141,26 +123,6 @@ const CreateRoomForm = ({ fields, onFormSubmit, onCreateFormReset }) => {
         }
     }
 
-    const validateEmail = (email) => {
-        const regexp = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-        return regexp.test(email)
-    }
-
-    const validatePhoneNumber = (phoneNumber) => {
-        let isValid = true
-
-        try {
-            parsePhoneNumber(phoneNumber, 'FR')
-        } catch (error) {
-            isValid = false
-        }
-
-        return isValid
-    }
-
-    const isParticipantValid = (participant) => {
-        return validateEmail(participant) || validatePhoneNumber(participant)
-    }
     interface MyInputProps {
         onKeyDown: (event: object) => void
     }
@@ -201,27 +163,6 @@ const CreateRoomForm = ({ fields, onFormSubmit, onCreateFormReset }) => {
                 {...getTagProps({ index })}
             />
         )
-    }
-
-    const isNumeric = (destination) => {
-        for (let char of destination) {
-            if (char >= '0' && char <= '9') continue
-            else return false
-        }
-
-        return true
-    }
-
-    const formatDestinations = (destinations) => {
-        return destinations.map((destination) => {
-            if (destination.includes('@')) {
-                return { email: destination }
-            } else if (isNumeric(destination)) {
-                return { phone: destination }
-            } else {
-                return { groupId: destination }
-            }
-        })
     }
 
     const cancelEdit = () => {
