+++ src/services/api.ts
@@ -4,7 +4,7 @@ import { NewRoomResponse } from '../../types/NewRoomResponse'
 import { RoomId } from '../../types/Room'
 import { UID } from '../../types/uid'
 import ApiClient from './apiClient'
-import { Room } from '../pages/AdminDashboard/CreateRoomForm'
+import { Room } from '../pages/AdminDashboard/CreateRoomForm/CreateRoomForm'
 
 export class Api {
     private apiClient: ApiClient
